Removed StoredTableDataset.
