+++ src/Pure.RelationalSchema.Storage/Pure.RelationalSchema.Storage.csproj
@@ -15,10 +15,7 @@
   <ItemGroup>
     <PackageReference Include="Pure.Primitives.Materialized" Version="0.1.0" />
     <PackageReference Include="Pure.Primitives.String.Operations" Version="0.2.2" />
-    <PackageReference
-      Include="Pure.RelationalSchema.Storage.Abstractions"
-      Version="0.1.0-preview.0.1.0"
-    />
+    <PackageReference Include="Pure.RelationalSchema.Storage.Abstractions" Version="0.1.0-preview.1.0.0" />
     <PackageReference Include="System.Linq.Expressions" Version="4.3.0" />
   </ItemGroup>
 </Project>
